requirements.txt
python-dotenv==1.0.0
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
sqlalchemy==2.0.23
pandas==2.1.3
scikit-learn==1.3.2 